updated test cases

+++ src/main/java/com/example/techolution/service/UserServiceImpl.java
@@ -1,9 +1,12 @@
 package com.example.techolution.service;
 
 import com.example.techolution.entity.User;
+import com.example.techolution.exception.AccessDeniedException;
 import com.example.techolution.exception.ResourceNotFoundException;
 import com.example.techolution.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -32,15 +35,17 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User addUser(User user) {
+        checkAuthorizationForRole("MANAGER", "ADMIN");
         return userRepository.save(user);
     }
 
     public User updateUser(Long userId, User user) {
+        checkAuthorizationForRole("MANAGER", "ADMIN");
         Optional<User> optionalUser = userRepository.findById(userId);
         if (optionalUser.isPresent()) {
             User existingUser = optionalUser.get();
             existingUser.setUserName(user.getUserName());
-            existingUser.setPasswordHash(user.getPasswordHash());
+            existingUser.setPassword(user.getPassword());
             return userRepository.save(existingUser);
         } else {
             throw new ResourceNotFoundException("User not found with ID: " + userId);
@@ -49,11 +54,23 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUserById(Long userId) {
+        checkAuthorizationForRole("ADMIN");
         if (userRepository.existsById(userId)) {
             userRepository.deleteById(userId);
         } else {
             throw new ResourceNotFoundException("User not found with ID: " + userId);
         }
+    }
+
+    private void checkAuthorizationForRole(String... allowedRoles) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        for (String allowedRole : allowedRoles) {
+            if (authentication.getAuthorities().stream().anyMatch(authority ->
+                    authority.getAuthority().equals("ROLE_" + allowedRole))) {
+                return;
+            }
+        }
+        throw new AccessDeniedException("Access denied. User does not have the required role.");
     }
 }
